Refactor chart plots file structure.
