+++ web/src/features/channel/charts/chartCanvas.ts
@@ -1,8 +1,7 @@
 import * as d3 from "d3";
 import { NumberValue, ScaleLinear, ScaleTime, Selection } from "d3";
 import { addHours, subHours } from "date-fns";
-import clone from "clone";
-import eventIcons from "./eventIcons";
+import { BarPlot } from "./plots/bar";
 
 type chartConfig = {
   leftYAxisKey: string;
@@ -26,7 +25,7 @@ type chartConfig = {
   xScale: ScaleTime<number, number, number | undefined>;
 };
 
-class Chart {
+class ChartCanvas {
   config: chartConfig = {
     xAxisPadding: 0,
     yAxisPadding: 1.2,
@@ -282,12 +281,9 @@ class Chart {
     return "rgb(" + ret.join(",") + ")";
   }
 
-  figures: Map<string, { plot: BarPlot; interactionConfig: any }> = new Map<
-    string,
-    { plot: BarPlot; interactionConfig: barInputConfig }
-  >();
+  figures: Map<string, { plot: BarPlot; drawConfig: any }> = new Map<string, { plot: BarPlot; drawConfig: any }>();
 
-  getFigure(xLocation: number, yLocation: number): { plot: BarPlot; interactionConfig: any } | undefined {
+  getFigure(xLocation: number, yLocation: number): { plot: BarPlot; drawConfig: any } | undefined {
     const colorData = this.interactionContext.getImageData(xLocation, yLocation, 1, 1).data;
 
     return this.figures.get("rgb(" + [colorData[0], colorData[1], colorData[2]].join(",") + ")");
@@ -491,468 +487,5 @@ class Chart {
     return this;
   }
 }
-type basePlotConfig = {
-  id: string; // The id used to fetch the Plot instance from the Chart instance
-  key: string; // The id used to fetch the Plot instance from the Chart instance
-};
 
-type drawConfig = {
-  xPosition?: number;
-  yPosition?: number;
-  xValue?: Date;
-  leftYValue?: number;
-  rightYValue?: number;
-  xIndex?: number;
-};
-
-abstract class AbstractPlot {
-  chart: Chart;
-  config: basePlotConfig;
-
-  constructor(chart: Chart, config: basePlotConfig) {
-    this.chart = chart;
-
-    this.config = {
-      ...config,
-    };
-  }
-
-  setConfig(config: basePlotConfig) {
-    this.config = { ...this.config, ...config };
-  }
-
-  getYScale() {
-    const yScaleMax = Math.max(
-      ...this.chart.data.map((d): number => {
-        return d[this.config.key];
-      })
-    );
-    return this.chart.config.yScale.copy().domain([yScaleMax * this.chart.config.yAxisPadding, 0]);
-  }
-
-  /**
-   * Used to offset the y pixel location effectively flipping the graph vertical axis,
-   * which is necessary because drawing has origin top left, while a chart/graph has origin bottom left.
-   */
-  offset(): number {
-    return (this.chart.canvas.node()?.height || 0) - this.chart.config.yScale.range()[1];
-  }
-
-  height(dataPoint: number): number {
-    return this.getYScale()(dataPoint);
-  }
-
-  /**
-   * xPoint returns the starting location for the bar on the xScale in pixels
-   *
-   * @param xValue the data point on the xScale that you want to convert to a pixel location on the chart.
-   */
-  xPoint(xValue: number): number {
-    return this.chart.config.xScale(xValue) || 0;
-  }
-
-  /**
-   * yPoint returns the starting location for the bar on the yScale in pixels
-   *
-   * @param yValue the data point on the yScale that you want to convert to a pixel location on the chart.
-   */
-  yPoint(yValue: number): number {
-    return this.height(yValue) + this.offset();
-  }
-
-  abstract draw(drawConfig: drawConfig): any;
-}
-
-type barsConfig = basePlotConfig & {
-  key: string; // The key used to fetch data
-  barGap: number; // The gap between each bar
-  barColor: string; // The color of the bar
-  barHoverColor: string;
-  labels?: boolean;
-  textColor: string;
-  textHoverColor: string;
-  cornerRadius: number; // The radius of the bar
-};
-type barInputConfig = Partial<barsConfig> & Required<Pick<barsConfig, "id" | "key">>;
-
-export class BarPlot extends AbstractPlot {
-  config: barsConfig;
-
-  /**
-   * Plots bars on a chart canvas. To use it add it to the plots map on the Chart instance.
-   *
-   * @param chart - The Chart instance where BarPlot will be plotted on
-   * @param config - Plot config, only required attributes are key and ID
-   */
-  constructor(chart: Chart, config: barInputConfig) {
-    super(chart, config);
-
-    this.config = {
-      barGap: 0.1,
-      barColor: "#B6DCFF",
-      barHoverColor: "#9DD0FF",
-      textColor: "#8198A3",
-      textHoverColor: "#3A463C",
-      cornerRadius: 3,
-      ...config,
-    };
-  }
-
-  /**
-   * xPoint returns the starting location for the bar on the xScale in pixels
-   *
-   * @param xValue the data point on the xScale that you want to convert to a pixel location on the chart.
-   */
-  xPoint(xValue: number): number {
-    return (this.chart.config.xScale(xValue) || 0) - this.barWidth() / 2;
-  }
-
-  barWidth(): number {
-    return (this.chart.config.xScale(new Date(1, 0, 1)) || 0) - (this.chart.config.xScale(new Date(1, 0, 0)) || 0);
-  }
-
-  drawBar(context: CanvasRenderingContext2D, dataPoint: any, fillColor: string) {
-    context.fillStyle = fillColor;
-    context.strokeStyle = fillColor;
-
-    // Draw the bar rectangle
-    context.fillRect(
-      this.xPoint(dataPoint.date) + this.config.cornerRadius / 2 + (this.barWidth() * this.config.barGap) / 2,
-      this.yPoint(dataPoint[this.config.key]) + this.config.cornerRadius / 2,
-      this.barWidth() * (1 - this.config.barGap) - this.config.cornerRadius,
-      this.height(-dataPoint[this.config.key]) - this.config.cornerRadius
-    );
-
-    // This draws the stroke used to create rounded corners
-    context.strokeRect(
-      this.xPoint(dataPoint.date) + this.config.cornerRadius / 2 + (this.barWidth() * this.config.barGap) / 2,
-      this.yPoint(dataPoint[this.config.key]) + this.config.cornerRadius / 2,
-      this.barWidth() * (1 - this.config.barGap) - this.config.cornerRadius,
-      this.height(-dataPoint[this.config.key]) - this.config.cornerRadius
-    );
-  }
-
-  /**
-   * Draw draws the bars on the Chart instance based on the configuration provided.
-   */
-  draw(drawConfig?: drawConfig) {
-    this.chart.data.forEach((data, i) => {
-      this.chart.context.fillStyle = this.config.barColor;
-      this.chart.context.strokeStyle = this.config.barColor;
-      this.chart.context.lineJoin = "round";
-      this.chart.context.lineWidth = this.config.cornerRadius;
-
-      let xIndex: number = -1;
-      if (
-        drawConfig?.xPosition &&
-        addHours(this.chart.config.xScale.invert(drawConfig.xPosition), 12) > data.date &&
-        addHours(this.chart.config.xScale.invert(drawConfig.xPosition), 12) < addHours(data.date, 24)
-      ) {
-        xIndex = i;
-      }
-
-      const hoversOverDataPoint = xIndex === i;
-
-      let barColor = this.config.barColor;
-      if (hoversOverDataPoint) {
-        barColor = this.config.barHoverColor;
-      }
-      this.drawBar(this.chart.context, this.chart.data[i], barColor);
-
-      // Create the interaction color and
-      const interactionColor = this.chart.genColor();
-      // this.chart.figures.set(interactionColor, {
-      //   plot: this,
-      //   interactionConfig: { hoverIndex: i },
-      // });
-
-      this.chart.interactionContext.fillStyle = interactionColor;
-      this.chart.interactionContext.fillRect(
-        Math.round(this.xPoint(this.chart.data[i].date)),
-        Math.round(this.yPoint(this.chart.data[i][this.config.key])),
-        Math.round(this.barWidth()),
-        Math.round(this.height(-this.chart.data[i][this.config.key]))
-      );
-    });
-    this.chart.data.forEach((data, i) => {
-      let xIndex: number = -1;
-      if (
-        drawConfig?.xPosition &&
-        addHours(this.chart.config.xScale.invert(drawConfig.xPosition), 12) > data.date &&
-        addHours(this.chart.config.xScale.invert(drawConfig.xPosition), 12) < addHours(data.date, 24)
-      ) {
-        xIndex = i;
-      }
-      const hoversOverDataPoint = xIndex === i;
-      let textColor = this.config.textColor;
-      if (hoversOverDataPoint) {
-        textColor = this.config.textHoverColor;
-      }
-
-      if (this.config.labels || hoversOverDataPoint) {
-        this.chart.context.font = "12px Inter";
-        this.chart.context.textAlign = "center";
-        this.chart.context.textBaseline = "middle";
-        this.chart.context.fillStyle = textColor;
-        this.chart.context.fillText(
-          d3.format(",")(this.chart.data[i][this.config.key]),
-          this.xPoint(this.chart.data[i].date) + this.barWidth() / 2,
-          this.yPoint(this.chart.data[i][this.config.key]) - 15 + this.config.cornerRadius / 2
-        );
-      }
-    });
-  }
-}
-
-type areaPlotConfig = basePlotConfig & {
-  areaColor: string;
-  areaGradient?: Array<string>[2];
-  addBuffer: boolean;
-  globalAlpha: number;
-  labels?: boolean;
-};
-
-type areaPlotConfigInit = Partial<areaPlotConfig> & basePlotConfig;
-
-export class AreaPlot extends AbstractPlot {
-  config: areaPlotConfig;
-  legend: Selection<HTMLDivElement, {}, HTMLElement, any>;
-  legendTextBox: Selection<HTMLDivElement, {}, HTMLElement, any>;
-  legendColorBox: Selection<HTMLDivElement, {}, HTMLElement, any>;
-
-  constructor(chart: Chart, config: areaPlotConfigInit) {
-    super(chart, config);
-
-    this.config = {
-      areaColor: "#DAEDFF",
-      globalAlpha: 1,
-      addBuffer: false,
-      ...config,
-    };
-
-    this.legend = this.chart.legendContainer
-      .append("div")
-      .attr("id", `${this.config.id}`)
-      .attr("style", `display: grid; grid-auto-flow: column; align-items: center; grid-column-gap: 5px;`);
-
-    this.legendTextBox = this.legend.append("div").attr("class", "legendTextBox");
-
-    const legendColor = this.config.areaGradient
-      ? `linear-gradient(0deg, ${this.config.areaGradient[0]} 0%, ${this.config.areaGradient[1]} 100%)`
-      : this.config.areaColor;
-
-    this.legendColorBox = this.legend
-      .append("div")
-      .attr("class", "legendColorBox")
-      .attr("style", `width: 12px; height: 12px; background: ${legendColor};`);
-  }
-
-  draw(drawConfig?: drawConfig) {
-    const area = d3
-      .area()
-      .x((d, i): number => {
-        return this.chart.config.xScale(this.chart.data[i].date) || 0;
-      })
-      .y0((d, i): number => {
-        return this.getYScale()(this.chart.data[i][this.config.key]) || 0;
-      })
-      .y1((d, i): number => {
-        return this.getYScale()(0) || 1;
-      })
-      .context(this.chart.context);
-
-    this.chart.context.globalAlpha = this.config.globalAlpha;
-
-    this.chart.context.fillStyle = this.config.areaColor;
-
-    if (this.config.areaGradient) {
-      let gradient = this.chart.context.createLinearGradient(0, 0, 0, this.chart.config.yScale.range()[1] || 0);
-      gradient.addColorStop(0, this.config.areaGradient[1]);
-      gradient.addColorStop(1, this.config.areaGradient[0]);
-      this.chart.context.fillStyle = gradient;
-    }
-
-    let data = this.chart.data;
-
-    if (data && this.config.addBuffer) {
-      let lastItem = clone(data[data.length - 1]);
-      let firstItem = clone(data[0]);
-      lastItem.date = addHours(lastItem.date, 16);
-      data.push(lastItem);
-      data.unshift(firstItem);
-      firstItem.date = subHours(firstItem.date, 16);
-    }
-
-    this.chart.context.beginPath();
-    area(data);
-
-    this.chart.context.fill();
-    this.chart.context.globalAlpha = 1;
-    if (data && this.config.addBuffer) {
-      data.splice(data.length - 1, 1);
-      data.splice(0, 1);
-    }
-
-    this.chart.data.forEach((d, i) => {
-      if (this.config.labels) {
-        this.chart.context.font = "12px Inter";
-        this.chart.context.textAlign = "center";
-        this.chart.context.textBaseline = "middle";
-        this.chart.context.fillStyle = "#3A463C";
-        this.chart.context.fillText(
-          d3.format(",")(d[this.config.key]),
-          this.xPoint(d.date),
-          this.yPoint(d[this.config.key]) - 15
-        );
-      }
-    });
-
-    const legendText = drawConfig?.xIndex
-      ? this.chart.data[drawConfig?.xIndex][this.config.key]
-      : this.chart.data[this.chart.data.length - 1][this.config.key];
-    this.legendTextBox.text(d3.format(",")(legendText));
-  }
-}
-
-type linePlotConfig = basePlotConfig & {
-  lineColor: string;
-  globalAlpha: number;
-  labels?: boolean;
-};
-
-type linePlotConfigInit = Partial<linePlotConfig> & basePlotConfig;
-
-export class LinePlot extends AbstractPlot {
-  config: linePlotConfig;
-  legend: Selection<HTMLDivElement, {}, HTMLElement, any>;
-  legendTextBox: Selection<HTMLDivElement, {}, HTMLElement, any>;
-  legendColorBox: Selection<HTMLDivElement, {}, HTMLElement, any>;
-
-  constructor(chart: Chart, config: linePlotConfigInit) {
-    super(chart, config);
-
-    this.config = {
-      lineColor: "#85C4FF",
-      globalAlpha: 1,
-      ...config,
-    };
-
-    this.legend = this.chart.legendContainer
-      .append("div")
-      .attr("id", `${this.config.id}`)
-      .attr(
-        "style",
-        `display: grid; grid-auto-flow: column; align-items: center; grid-column-gap: 5px; justify-content: end;`
-      );
-
-    this.legendTextBox = this.legend.append("div").attr("class", "legendTextBox");
-
-    this.legendColorBox = this.legend
-      .append("div")
-      .attr("class", "legendColorBox")
-      .attr("style", `width: 12px; height: 12px; background: ${this.config.lineColor};`);
-  }
-
-  draw(drawConfig?: drawConfig) {
-    const line = d3
-      .line()
-      .x((d, i): number => {
-        return this.chart.config.xScale(this.chart.data[i].date) || 0;
-      })
-      .y((d, i): number => {
-        return this.getYScale()(this.chart.data[i][this.config.key]) || 0;
-      })
-      .context(this.chart.context);
-
-    this.chart.context.globalAlpha = this.config.globalAlpha;
-    this.chart.context.strokeStyle = this.config.lineColor;
-
-    this.chart.context.beginPath();
-    line(this.chart.data);
-
-    this.chart.context.stroke();
-    this.chart.context.globalAlpha = 1;
-
-    if (this.config.labels) {
-      this.chart.data.forEach((d, i) => {
-        this.chart.context.font = "12px Inter";
-        this.chart.context.textAlign = "center";
-        this.chart.context.textBaseline = "middle";
-        this.chart.context.fillStyle = "#3A463C";
-        this.chart.context.fillText(
-          d3.format(",")(d[this.config.key]),
-          this.xPoint(d.date),
-          this.yPoint(d[this.config.key]) - 15
-        );
-      });
-    }
-
-    const hoverIndex = drawConfig?.xIndex || this.chart.data.length - 1;
-    const legendText = this.chart.data[hoverIndex][this.config.key];
-    this.legendTextBox.text(d3.format(",")(legendText));
-  }
-}
-
-type eventsPlotConfig = basePlotConfig;
-
-export class EventsPlot extends AbstractPlot {
-  config: eventsPlotConfig;
-
-  lastWidth?: number;
-  lastHeight?: number;
-
-  constructor(chart: Chart, config: eventsPlotConfig) {
-    super(chart, config);
-
-    this.config = {
-      ...config,
-    };
-  }
-
-  draw(drawConfig?: drawConfig) {
-    // Don't redraw events if nothing has changed as it is expensive to add/remove html elements
-    if (this.lastWidth === this.chart.config.width && this.lastHeight === this.chart.config.height) {
-      return;
-    }
-    this.lastWidth = this.chart.config.width;
-    this.lastHeight = this.chart.config.height;
-
-    // Clear events
-    this.chart.eventsContainer.selectAll("*").remove();
-
-    this.chart.eventsContainer
-      .selectAll(".event-wrapper")
-      .data(this.chart.data)
-      .enter()
-      .append("div")
-      .attr("class", "event-wrapper")
-      .attr("id", function (d: any, i) {
-        return "event-" + i;
-      })
-      .attr("style", (d: any, i) => {
-        return `position:absolute; left: ${this.xPoint(d.date)}px; bottom:5px;`;
-      })
-      .selectAll(".event-item")
-      .data((d: any, i: number) => {
-        return d.events || [];
-      })
-      .enter()
-      .append("div")
-      .attr("class", (d: any, o) => {
-        return "event-item " + d.type;
-      })
-      .html((d: any, i) => {
-        const icon = eventIcons.get(d.type) || "";
-        const text = d3.format(".2s")(d.value);
-        if (d.value === undefined) {
-          return icon;
-        } else if (d.value === 0) {
-          return "0" + icon;
-        } else {
-          return `<span class="event-item-text">${text}</span>` + icon;
-        }
-      });
-  }
-}
-
-export default Chart;
+export default ChartCanvas;
