+++ web/src/features/channel/charts/charts.ts
@@ -1,0 +1,5 @@
+export { default as ChartCanvas } from "./chartCanvas";
+export { AreaPlot } from "./plots/area";
+export { LinePlot } from "./plots/line";
+export { BarPlot } from "./plots/bar";
+export { EventsPlot } from "./plots/events";
