+++ web/src/features/channel/charts/plots/eventIcons.ts
