+++ web/src/features/channel/revenueChart/ProtifsChart.tsx
@@ -2,7 +2,8 @@
 import { useD3 } from "../charts/useD3";
 import React, { useEffect } from "react";
 import { Selection } from "d3";
-import Chart, { BarPlot } from "../charts/chart";
+import ChartCanvas from "../charts/chartCanvas";
+import { BarPlot } from "../charts/plots/bar";
 import "../charts/chart.scss";
 
 type ProfitsChart = {
@@ -10,12 +11,12 @@ type ProfitsChart = {
 };
 
 function ProfitsChart({ data }: ProfitsChart) {
-  let chart: Chart;
+  let chart: ChartCanvas;
 
   // TODO: Change this so that we can update the data without redrawing the entire chart
   const ref = useD3(
     (container: Selection<HTMLDivElement, {}, HTMLElement, any>) => {
-      chart = new Chart(container, data, { leftYAxisKey: "revenue", xAxisPadding: 12 });
+      chart = new ChartCanvas(container, data, { leftYAxisKey: "revenue", xAxisPadding: 12 });
       chart.plot(BarPlot, { id: "bars", key: "revenue" });
       chart.draw();
     },
